refactored and moved the RunnerResources class
