+++ opuspocus/runners/slurm.py
@@ -11,8 +11,9 @@ from attrs import Attribute, define, field, validators
 
 from opuspocus.pipeline_steps import OpusPocusStep
 from opuspocus.pipelines import OpusPocusPipeline
+from opuspocus.runner_resources import RunnerResources
 from opuspocus.runners import OpusPocusRunner, TaskInfo, register_runner
-from opuspocus.utils import RunnerResources, subprocess_wait
+from opuspocus.utils import subprocess_wait
 
 logger = logging.getLogger(__name__)
